requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.3
openai==1.13.3
redis==5.0.3
python-dotenv==1.0.1
httpx==0.27.2